requirements.txt
contourpy==1.3.2
cycler==0.12.1
Faker==37.8.0
fonttools==4.60.0
kiwisolver==1.4.9
matplotlib==3.10.6
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pycryptodome==3.23.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
